+++ src/CacheProfiles/CacheAllSuccessfulGetRequests.php
@@ -2,6 +2,7 @@
 
 namespace Spatie\ResponseCache\CacheProfiles;
 
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -22,6 +23,34 @@ class CacheAllSuccessfulGetRequests extends BaseCacheProfile
 
     public function shouldCacheResponse(Response $response): bool
     {
-        return $response->isSuccessful() || $response->isRedirection();
+        if (! $this->hasCacheableResponseCode($response)) {
+            return false;
+        }
+
+        if (! $this->hasCacheableContentType($response)) {
+            return false;
+        }
+
+        return true;
+    }
+
+    public function hasCacheableResponseCode(Response $response): bool
+    {
+        if ($response->isSuccessful()) {
+            return true;
+        }
+
+        if ($response->isRedirection()) {
+            return true;
+        }
+
+        return false;
+    }
+
+    public function hasCacheableContentType(Response $response)
+    {
+        $contentType = $response->headers->get('Content-Type', '');
+
+        return Str::startsWith($contentType, 'text');
     }
 }