+++ src/CacheProfiles/CacheProfile.php
@@ -8,24 +8,12 @@ use Symfony\Component\HttpFoundation\Response;
 
 interface CacheProfile
 {
-    /*
-     * Determine if the response cache middleware should be enabled.
-     */
     public function enabled(Request $request): bool;
 
-    /*
-     * Determine if the given request should be cached.
-     */
     public function shouldCacheRequest(Request $request): bool;
 
-    /*
-     * Determine if the given response should be cached.
-     */
     public function shouldCacheResponse(Response $response): bool;
 
-    /*
-     * Return the time when the cache must be invalidated.
-     */
     public function cacheRequestUntil(Request $request): DateTime;
 
     /*
@@ -34,5 +22,5 @@ interface CacheProfile
      * For example: if you want a different cache per user you could return the id of
      * the logged in user.
      */
-    public function cacheNameSuffix(Request $request): string;
+    public function useCacheNameSuffix(Request $request): string;
 }