+++ src/Facades/ResponseCache.php
@@ -6,13 +6,6 @@ use Illuminate\Support\Facades\Facade;
 
 class ResponseCache extends Facade
 {
-    /**
-     * Get the registered name of the component.
-     *
-     * @see \Spatie\ResponseCache\ResponseCache
-     *
-     * @return string
-     */
     protected static function getFacadeAccessor()
     {
         return 'responsecache';