+++ src/Hasher/RequestHasher.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Spatie\ResponseCache\Hasher;
+
+use Illuminate\Http\Request;
+
+interface RequestHasher
+{
+    public function getHashFor(Request $request): string;
+}