+++ src/Middlewares/CacheResponse.php
@@ -4,8 +4,10 @@ namespace Spatie\ResponseCache\Middlewares;
 
 use Closure;
 use Illuminate\Http\Request;
+use Illuminate\Support\Collection;
 use Spatie\ResponseCache\ResponseCache;
 use Spatie\ResponseCache\Events\CacheMissed;
+use Spatie\ResponseCache\Replacers\Replacer;
 use Symfony\Component\HttpFoundation\Response;
 use Spatie\ResponseCache\Events\ResponseCacheHit;
 
@@ -25,7 +27,13 @@ class CacheResponse
             if ($this->responseCache->hasBeenCached($request)) {
                 event(new ResponseCacheHit($request));
 
-                return $this->responseCache->getCachedResponseFor($request);
+                $response = $this->responseCache->getCachedResponseFor($request);
+
+                $this->getReplacers()->each(function (Replacer $replacer) use ($response) {
+                    $replacer->replaceInCachedResponse($response);
+                });
+
+                return $response;
             }
         }
 
@@ -33,12 +41,34 @@ class CacheResponse
 
         if ($this->responseCache->enabled($request)) {
             if ($this->responseCache->shouldCache($request, $response)) {
-                $this->responseCache->cacheResponse($request, $response, $lifetimeInSeconds);
+                $this->makeReplacementsAndCacheResponse($request, $response, $lifetimeInSeconds);
             }
         }
 
         event(new CacheMissed($request));
 
         return $response;
+    }
+
+    protected function makeReplacementsAndCacheResponse(
+        Request $request,
+        Response $response,
+        $lifetimeInSeconds = null
+    ): void {
+        $cachedResponse = clone $response;
+
+        $this->getReplacers()->each(function (Replacer $replacer) use ($cachedResponse) {
+            $replacer->prepareResponseToCache($cachedResponse);
+        });
+
+        $this->responseCache->cacheResponse($request, $cachedResponse, $lifetimeInSeconds);
+    }
+
+    protected function getReplacers(): Collection
+    {
+        return collect(config('responsecache.replacers'))
+            ->map(function (string $replacerClass) {
+                return app($replacerClass);
+            });
     }
 }