+++ src/Middlewares/DoNotCacheResponse.php
@@ -7,14 +7,6 @@ use Illuminate\Http\Request;
 
 class DoNotCacheResponse
 {
-    /**
-     * Handle an incoming request.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @param \Closure $next
-     *
-     * @return mixed
-     */
     public function handle(Request $request, Closure $next)
     {
         $request->attributes->add(['responsecache.doNotCache' => true]);