+++ src/Replacers/CsrfTokenReplacer.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Spatie\ResponseCache\Replacers;
+
+use Symfony\Component\HttpFoundation\Response;
+
+class CsrfTokenReplacer implements Replacer
+{
+    protected $replacementString = '<laravel-responsecache-csrf-token-here>';
+
+    public function prepareResponseToCache(Response $response): void
+    {
+        if (! $response->getContent()) {
+            return;
+        }
+
+        $response->setContent(str_replace(
+            csrf_token(),
+            $this->replacementString,
+            $response->getContent()
+        ));
+    }
+
+    public function replaceInCachedResponse(Response $response): void
+    {
+        if (! $response->getContent()) {
+            return;
+        }
+
+        $response->setContent(str_replace(
+            $this->replacementString,
+            csrf_token(),
+            $response->getContent()
+        ));
+    }
+}