+++ src/ResponseCache.php
@@ -3,6 +3,8 @@
 namespace Spatie\ResponseCache;
 
 use Illuminate\Http\Request;
+use Spatie\ResponseCache\Hasher\DefaultHasher;
+use Spatie\ResponseCache\Hasher\RequestHasher;
 use Symfony\Component\HttpFoundation\Response;
 use Spatie\ResponseCache\CacheProfiles\CacheProfile;
 
@@ -11,7 +13,7 @@ class ResponseCache
     /** @var \Spatie\ResponseCache\ResponseCache */
     protected $cache;
 
-    /** @var \Spatie\ResponseCache\RequestHasher */
+    /** @var RequestHasher */
     protected $hasher;
 
     /** @var \Spatie\ResponseCache\CacheProfiles\CacheProfile */
@@ -48,10 +50,14 @@ class ResponseCache
             $response = $this->addCachedHeader($response);
         }
 
+        $lifetimeInSeconds = $lifetimeInSeconds
+            ? (int)$lifetimeInSeconds
+            : $this->cacheProfile->cacheRequestUntil($request);
+
         $this->cache->put(
             $this->hasher->getHashFor($request),
             $response,
-            ($lifetimeInSeconds) ? intval($lifetimeInSeconds) : $this->cacheProfile->cacheRequestUntil($request)
+            $lifetimeInSeconds,
         );
 
         return $response;
@@ -67,14 +73,6 @@ class ResponseCache
     public function getCachedResponseFor(Request $request): Response
     {
         return $this->cache->get($this->hasher->getHashFor($request));
-    }
-
-    /**
-     * @deprecated Use the new clear method, this is just an alias.
-     */
-    public function flush()
-    {
-        $this->clear();
     }
 
     public function clear()