+++ tests/CacheProfiles/CacheAllSuccessfulGetRequestsTest.php
@@ -2,7 +2,6 @@
 
 namespace Spatie\ResponseCache\Test\CacheProfiles;
 
-use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Spatie\ResponseCache\Test\User;
 use Spatie\ResponseCache\Test\TestCase;
@@ -44,6 +43,16 @@ class CacheAllSuccessfulGetRequestsTest extends TestCase
     }
 
     /** @test */
+    public function it_will_determine_that_a_non_text_response_should_not_be_cached()
+    {
+        $response = $this->createResponse(200, 'application/pdf');
+
+        $shouldCacheResponse = $this->cacheProfile->shouldCacheResponse($response);
+
+        $this->assertFalse($shouldCacheResponse);
+    }
+
+    /** @test */
     public function it_will_determine_that_an_error_should_not_be_cached()
     {
         foreach (range(400, 599) as $statusCode) {
@@ -54,18 +63,18 @@ class CacheAllSuccessfulGetRequestsTest extends TestCase
     /** @test */
     public function it_will_use_the_id_of_the_logged_in_user_to_differentiate_caches()
     {
-        $this->assertEquals('', $this->cacheProfile->cacheNameSuffix($this->createRequest('get')));
+        $this->assertEquals('', $this->cacheProfile->useCacheNameSuffix($this->createRequest('get')));
 
         User::all()->map(function ($user) {
             auth()->login(User::find($user->id));
-            $this->assertEquals($user->id, $this->cacheProfile->cacheNameSuffix($this->createRequest('get')));
+            $this->assertEquals($user->id, $this->cacheProfile->useCacheNameSuffix($this->createRequest('get')));
         });
     }
 
     /** @test */
     public function it_will_determine_to_cache_responses_for_a_certain_amount_of_time()
     {
-        /** @var $expirationDate Carbon */
+        /** @var $expirationDate \Carbon\Carbon */
         $expirationDate = $this->cacheProfile->cacheRequestUntil($this->createRequest('get'));
 
         $this->assertTrue($expirationDate->isFuture());
@@ -94,11 +103,13 @@ class CacheAllSuccessfulGetRequestsTest extends TestCase
      *
      * @return \Symfony\Component\HttpFoundation\Response;
      */
-    protected function createResponse($statusCode)
+    protected function createResponse($statusCode, $contentType = 'text/html; charset=UTF-8')
     {
         $response = new Response();
 
-        $response->setStatusCode($statusCode);
+        $response
+            ->setStatusCode($statusCode)
+            ->headers->set('Content-Type', $contentType);
 
         return $response;
     }