+++ tests/User.php
@@ -10,10 +10,5 @@ class User extends Model implements AuthenticatableContract
 {
     use Authenticatable;
 
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array
-     */
     protected $fillable = ['name', 'email', 'password'];
 }